api/serializers.py
from rest_framework import serializers
from django.db.models import User

class LoginSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ['username', 'email', 'password']