requirements.txt
aniso8601==7.0.0
asgiref==3.5.2
Django==4.1.10
django-annoying==0.10.6
django-cors-headers==3.5.0
django-crontab==0.7.1
django-extensions==3.1.5
django-filter==23.2
django-jsonstore==0.5.1
django-pandas==0.6.6
django-phonenumber-field==6.3.0
django-smart-pagination==1.0.1
django-viewflow==1.6.1
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.2
pandas==1.4.3
pendulum==2.1.2
phonenumbers==8.12.54
Pillow==9.2.0
promise==2.3
psycopg2-binary==2.9.3
python-dateutil==2.8.2
pytz==2022.2.1
pytzdata==2020.1
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
